% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_content.R
\name{solution}
\alias{solution}
\alias{solution_begin}
\alias{solution_end}
\title{Indicate a "solution" within the book.}
\usage{
solution(text, header = "&#x25BA; Solution", toggle = TRUE, hidden = FALSE)

solution_begin(header = "&#x25BA; Solution", toggle = TRUE, hidden = FALSE)

solution_end()
}
\arguments{
\item{text}{\code{character}. Text that provides the body of the solution.}

\item{header}{\code{character}. Defines the text used to mark the start
of a 'solution'.  Set this if you want to refer to this as an 'Answer',
'Explanation', etc.
'&#x25BA;' renders as a filled right pointing triangle in HTML.}

\item{toggle}{\code{logical}.  If set to \code{TRUE} the solution is
rendered in a collapsed format, with only the header defined in \code{header}
visible and a '+' icon that will expand the solution when clicked. Setting
to \code{FALSE} renders the complete solution with no option to collapse it.
Default value is \code{TRUE}.}

\item{hidden}{\code{logical}.  If set to \code{TRUE} the solution will be
included in the HTML, but it will not be visible i.e. \code{display: none;}.
This allow solutions to be generated during rendering, but they are
completely hidden from the user in cases where the intention is to provide
them at a later date.  Default value is \code{FALSE}.}
}
\description{
Functions to demark a 'solution' within an R Markdown document.
}
\details{
If your 'solution' is a single paragraph of text, you can use
  \code{solution()} and provide the paragraph via the argument \code{text}.

If you need to provide more detail than a single paragraph,
  you can use the functions \code{solution_begin()} and \code{solution_end()}
  to indicated the region of the document you wish to be included in the
  question.  (This is similar to using \code{\\begin{}} and \code{\\end{}}
  tags in a Latex document.).  These two functions must be used together,
  otherwise the resulting document will have unmatched HTML tags.

\code{solution_begin()} is a simple wrapper around \code{solution} with the
  \code{text} forced to be missing, and you can start a larger solution with
  either.  \code{solution_end()} should never be used on its own as it
  closes bits of document structure that are expected to have been opened
  already.  Doing so will have unexpected results on the final layout.
}
\seealso{
\code{\link{question}}
}
