% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_resources}
\alias{template_resources}
\title{change template ressources if required}
\usage{
template_resources(name, package, ...)
}
\description{
modified version of tufte:::template_resources that allows
us to specify the package where the resource can be found.
rather than being hardwired to "tufte"
}
