% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toggle_script}
\alias{toggle_script}
\title{javascript to be included for toggling solution visiblity}
\usage{
toggle_script()
}
\description{
There must be a more elegant way of including this,
 but it'll do for now.
}
