% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_templ_book.R
\name{unhcr_templ_book}
\alias{unhcr_templ_book}
\title{UNHCR Book template for bookdown

Loads additional style and template file}
\usage{
unhcr_templ_book(..., margin_references = TRUE)
}
\arguments{
\item{margin_references}{\code{logical}.  Determines whether to place
citations in the margin, or collate them at the end of the document.}
}
\description{
UNHCR Book template for bookdown

Loads additional style and template file
}
\details{
\code{unhcr_templ_book()} is used to create output split
into individual chapters for a book.  If you wish to create a complete
book this should be used in the output section of your YAML header
in the R Markdown document.
}
\references{
https://github.com/grimbough/msmbstyle
}
